rename penalize() to report_bad_message() and move DownloadError
