+++ crates/interfaces/src/test_utils/headers.rs
@@ -2,8 +2,8 @@
 use crate::{
     consensus::{self, Consensus},
     p2p::{
-        downloader::{DownloadClient, DownloadError, DownloadResult, DownloadStream, Downloader},
-        error::{PeerRequestResult, RequestError},
+        downloader::{DownloadClient, DownloadStream, Downloader},
+        error::{DownloadError, DownloadResult, PeerRequestResult, RequestError},
         headers::{
             client::{HeadersClient, HeadersRequest},
             downloader::HeaderDownloader,
@@ -168,7 +168,7 @@ impl TestHeadersClient {
 }
 
 impl DownloadClient for TestHeadersClient {
-    fn penalize(&self, _peer_id: PeerId) {
+    fn report_bad_message(&self, _peer_id: PeerId) {
         // noop
     }
 }
